fix: [한병찬] fix request form data type

+++ user_spent_predict/controller/request_form/user_spent_predict_request_form.py
@@ -5,8 +5,8 @@ class UserSpentPredictRequestForm(BaseModel):
     gender: str
     birth_year: int
     num_logins: int
-    average_login_interval: int
+    average_login_interval: float
     days_from_last_login: int
     member_maintenance: int
     num_orders: int
-    average_order_interval: int
+    average_order_interval: float
